+++ GamePlay/ECS/KERenderComponent.cpp
@@ -2,5 +2,5 @@
 
 
 
-uint32_t KERenderComponent::total_vertex_num = 0;
-uint32_t KERenderComponent::total_index_num = 0;
+uint64_t KERenderComponent::total_vertex_num = 0;
+uint64_t KERenderComponent::total_index_num = 0;
